Make traverse method general purpose (#50)
Instead of taking a sequence where the first element is the
document origin (as used in the inliner), the sequence is now
just the path in the document. (Also, technically it's now an
iterable as it doesn't need to support indexing.)

For the inliner code, it's trivial to split the path parts from
the document origin. this should make the `traverse` method more
general purpose (e.g. for json schema normalisation).