RefInliner includes whole documents (#56)
This is inefficient when only one item from a large schema is inlined, but this way inlined schemas are still valid (my previous way of flattening refs included `/` in the property names). Schemas are inlined under a reserved "remote" property at the same level of "definitions" and "properties". As a side effect, RefRenamer no longer has to worry about clobbering keys, so that code was simplified.

Had to fix example schemas - need to do more research on why this wasn't caught earlier (#66).