Flattener wasn't using integers for list indices
The issue is that because `fragment_decode` does not know about
the structure of the document, numbers cannot automatically be
converted to integers, as e.g. "0" is a valid key in an object.

The solution I've used is to have `traverse` return the correct
path, since it "knows" when it hits an array.