Corrected use of json-pointer per RFC6901
After onboarding to a stricter validation tool than we've been using thus far
(https://github.com/everit-org/json-schema) I discovered that we've been
mis-using the `json-pointer` format for string types.

According to the RFC6901, a json-pointer is only valid when each token or
fragment is prefixed with a `/` character, hence leading with a `#` is invalid.
We'd been incorreclty using `#` to indicate document root in a json-pointer,
but the RFC actually indicates this format is only valid for encoding a
json-pointer inside a URI fragment. So, a value like "#/properties/propertyA"
is valid as the value for a `$ref`, for which the type is uri-reference but
the json-pointer itself implies document root in "/properties/propertyA"

This impacts how we were exemplifying our semantic schema enhancements
such as `createOnly` and `readOnly` and so I have corrected those examples.