Renaming `readOnly` and `writeOnly` block
After further study of the JSON Schema RFC, in particular section 10.3
related to validation annotations, I discovered that the `readOnly` and
`writeOnly` keywords are part of the schema validation vocabulary and
cannot be re-purposed as schema property keywords in the way we had done.

As such, this commit renames the semantic keywords to something without
conflicts in the JSON Schema vocabulary, and updates the README and examples
to reflect this change.

SEE: https://json-schema.org/draft-07/json-schema-validation.html#rfc.section.10.3