Codegen and wrapper wip (#178)
* Renamed `BaseHandler` to `HandlerWrapper` and refactored

`Base` indicates a type that can (or should) be extended and that's not the case here.
I also have a need to create a `BaseHandler` class for combining shared handler logic such as SDK Client initialisation.

* Added Configuration modelling to codegen

Configuration is used (currently) by handlers to produce their schema, which should be packaged in. The schema is used to validate incoming models for handler requests.