Correct IAM Policy for CloudWatch Logs
The IAM permission for CloudWatch Logs was pointing at a LogGroup created by the HandlerInfrastructure, rather than the one which LambdaLogger tries to write to by default. Shuffled the resources around to correctly setup an IAM policy which allows Handler functions to write to CloudWatch Logs. At this point, the LogGroup resource in the CloudFormationHandlerInfrastructure.yaml is somewhat superfluous but haven't removed it yet.