Adds logging infra and includes LoggingConfig (#276)
An IAM::Role with required CloudWatch permissions is now included in the managed-upload-infrastructure stack and the ARN of the role is passed to the CloudFormation RegisterType API as part of `cfn-cli submit`. A log group name is created for the type (e.g; initech-tps-report-logs) as part of the API.

I have verified in us-west-2 with the out-of-box type that I am now receiving logs to my own account from the published types.