Update delete contract tests to use output from create
This change updates the delete contract tests to extract the model from
completed creation tests and generate a request to delete that resource.

This change was successfully tested with a custom resource that creates
an EC2 instance and updates its model with the ID of the created
instance.

This change additionally includes some small refactoring to consolidate
repeated logic.