Increase timeout for local Lambda client (#318)
Java only: Some JAR files become so bloated, it takes SAM a long time to mount them inside the container. This happens on every request.

There is also a timeout value in the SAM template, this can be adjusted by the customer, so I'm leaving it at one minute for now.