Removed empty model assert in contract tests
Removed an assert requiring that ProgressEvent results with a status of FAILED should not include a ResourceModel.

This requirement is not actually part of the handler contract: https://docs.aws.amazon.com/cloudformation-cli/latest/userguide/resource-type-test-contract.html and was likely added to ensure consistency on this result as the model is strictly ignored in this case.

However, given this is not a documented part of the contract and there are no adverse outcomes from including a model in FAILED events, the test is an impediment and not a protection.