Nicer CLI experience when asking user for inputs on the terminal
As documented [here](https://docs.python.org/3/library/functions.html#input), this mechanism provides a nicer experience when we ask the user for input. When the user uses the cursor keys, they can actually navigate back and forth when using left and right and can navigate the history using up and down keys.